[SDCISA-4731] Fix out-of-order issue
Fixes the out-of-order issue. But this comes with a cost. Now we do all the
requests strictly in sequence and cannot fetch ahead a few in parallel as I
would like. But there's a chance that we always are faster than our downstream
can consume. If this guess is correct, it is even better to keep them
sequencial and don't do any fetch-ahead as those only would flood our buffer.